requirements.txt
flask
tensorflow
pillow
python-dotenv
gunicorn
requests
numpy
